Scrolling added, friction and animations fixed
Scrolling has been added, it attempts to keep the play in the center of
the screen, but will also keep the camera with in the bounds of the
level.  Animations are fixed, you walk while pressing the button, and
having an x velocity.  Friction is applied properly so you no longer
move through walls.  Hitting a ceiling causes you to lose some x
velocity.  This can be changed.